\name{Example.Section.3.1.2.corn}
\alias{Example.Section.3.1.2.corn}
\title{Example in Section.3.1.2 of Li (2019)}
\description{This is the example in Section.3.1.2 of Li (2019), which is a Leontief-type two-sector corn economy.}
\usage{Example.Section.3.1.2.corn()}
\author{
LI Wu <liwu@staff.shu.edu.cn>
}
\references{
LI Wu (2019, ISBN: 9787521804225) General Equilibrium and Structural Dynamics: Perspectives of New Structural Economics. Beijing: Economic Science Press. (In Chinese)
}
