\name{Example.Varian.P352}
\alias{Example.Varian.P352}
\title{Example on Page 352 in Varian (1992)}
\description{This is the example on page 352 in Varian (1992) (see also Example 15.C.2. in MWG, 1995, P542), which is a decreasing-returns-to-scale Cobb-Douglas 3-by-2 economy and can be transformed into a constant-returns-to-scale 3-by-3 (or 3-by-2) economy.}
\usage{
Example.Varian.P352(agent.number = 3)
}
\arguments{
\item{agent.number}{agent.number can be set to 3 or 2.}
}
\author{
LI Wu <liwu@staff.shu.edu.cn>
}
\references{
LI Wu (2019, ISBN: 9787521804225) General Equilibrium and Structural Dynamics: Perspectives of New Structural Economics. Beijing: Economic Science Press. (In Chinese)

Mas-Colell, Andreu and Whinston, Michael Dennis and Green, Jerry R. (1995, ISBN: 0195073401) Microeconomic Theory. Oxford University Press (New York).

Varian, Hal R. (1992, ISBN: 0393957357) Microeconomic Analysis. W. W. Norton & Company.
}
\examples{
Example.Varian.P352()

#####
Example.Varian.P352(agent.number = 2)
}
