\name{Example2.3}
\alias{Example2.3}
\title{Example2.3 in Li (2019)}
\description{This is Example 2.3 in Li (2019), which is a von Neumann economy.}
\usage{Example2.3()}
\author{
LI Wu <liwu@staff.shu.edu.cn>
}
\references{
LI Wu (2019, ISBN: 9787521804225) General Equilibrium and Structural Dynamics: Perspectives of New Structural Economics. Beijing: Economic Science Press. (In Chinese)
}
