\name{Example3.1}
\alias{Example3.1}
\title{Example 3.1 in Li (2019)}
\description{This is Example 3.1 in Li (2019),which is a two-sector corn economy with a non-homothetic utility function.}
\usage{
Example3.1()
}
\author{
LI Wu <liwu@staff.shu.edu.cn>
}
\references{
LI Wu (2019, ISBN: 9787521804225) General Equilibrium and Structural Dynamics: Perspectives of New Structural Economics. Beijing: Economic Science Press. (In Chinese)
}
