\name{Example3.10}
\alias{Example3.10}
\title{Example 3.10 in Li (2019)}
\description{This is Example 3.10 in Li (2019),which is a Leontief corn economy with three primary factors.}
\usage{Example3.10()}
\author{
LI Wu <liwu@staff.shu.edu.cn>
}
\references{
LI Wu (2019, ISBN: 9787521804225) General Equilibrium and Structural Dynamics: Perspectives of New Structural Economics. Beijing: Economic Science Press. (In Chinese)
}
