\name{Example3.12}
\alias{Example3.12}
\title{Example 3.12 in Li (2019)}
\description{This is Example 3.12 in Li (2019),which is an economy with decreasing returns to scale.}
\usage{Example3.12()}
\author{
LI Wu <liwu@staff.shu.edu.cn>
}
\references{
LI Wu (2019, ISBN: 9787521804225) General Equilibrium and Structural Dynamics: Perspectives of New Structural Economics. Beijing: Economic Science Press. (In Chinese)
}
