\name{Example3.14}
\alias{Example3.14}
\title{Example 3.14 in Li (2019)}
\description{This is Example 3.14 in Li (2019),which illustrates the relationship between a regular economy and a pure exchange economy.}
\usage{
Example3.14()
}
\author{
LI Wu <liwu@staff.shu.edu.cn>
}
\references{
LI Wu (2019, ISBN: 9787521804225) General Equilibrium and Structural Dynamics: Perspectives of New Structural Economics. Beijing: Economic Science Press. (In Chinese)
}
