\name{Example3.2}
\alias{Example3.2}
\title{Example 3.2 in Li (2019)}
\description{This is Example 3.2 in Li (2019),which is a Cobb-Douglas two-sector corn economy.}
\usage{Example3.2()}
\author{
LI Wu <liwu@staff.shu.edu.cn>
}
\references{
LI Wu (2019, ISBN: 9787521804225) General Equilibrium and Structural Dynamics: Perspectives of New Structural Economics. Beijing: Economic Science Press. (In Chinese)
}
