\name{Example3.4}
\alias{Example3.4}
\title{Example 3.4 in Li (2019)}
\description{This is Example 3.2 in Li (2019),which is a Lontief three-sector economy with one primary factor.}
\usage{
Example3.4()
}
\author{
LI Wu <liwu@staff.shu.edu.cn>
}
\references{
LI Wu (2019, ISBN: 9787521804225) General Equilibrium and Structural Dynamics: Perspectives of New Structural Economics. Beijing: Economic Science Press. (In Chinese)
}
