\name{Example3.8}
\alias{Example3.8}
\title{Example 3.8 in Li (2019)}
\description{This is Example 3.8 in Li (2019),which is a Cobb-Douglas three-sector economy with one primary factor.}
\usage{
Example3.8()
}
\author{
LI Wu <liwu@staff.shu.edu.cn>
}
\references{
LI Wu (2019, ISBN: 9787521804225) General Equilibrium and Structural Dynamics: Perspectives of New Structural Economics. Beijing: Economic Science Press. (In Chinese)
}
