\name{Example3.9}
\alias{Example3.9}
\title{Example 3.9 in Li (2019)}
\description{This is Example 3.9 in Li (2019),which is a Cobb-Douglas three-sector economy with two primary factors.}
\usage{
Example3.9()
}
\author{
LI Wu <liwu@staff.shu.edu.cn>
}
\references{
LI Wu (2019, ISBN: 9787521804225) General Equilibrium and Structural Dynamics: Perspectives of New Structural Economics. Beijing: Economic Science Press. (In Chinese)
}
