\name{Example4.11.2}
\alias{Example4.11.2}
\title{Second Part of Example 4.11 in Li (2019)}
\description{This is the second part of Example 4.11 in Li (2019),which illustrates the tax.}
\usage{Example4.11.2()}
\author{
LI Wu <liwu@staff.shu.edu.cn>
}
\references{
LI Wu (2019, ISBN: 9787521804225) General Equilibrium and Structural Dynamics: Perspectives of New Structural Economics. Beijing: Economic Science Press. (In Chinese)
}
