\name{Example4.13}
\alias{Example4.13}
\title{Example 4.13 in Li (2019)}
\description{This is Example 4.13 in Li (2019),which illustrates the divident.}
\usage{Example4.13()}
\author{
LI Wu <liwu@staff.shu.edu.cn>
}
\references{
LI Wu (2019, ISBN: 9787521804225) General Equilibrium and Structural Dynamics: Perspectives of New Structural Economics. Beijing: Economic Science Press. (In Chinese)
}
