\name{Example4.2}
\alias{Example4.2}
\title{Example 4.2 in Li (2019)}
\description{This is Example 4.2 in Li (2019), which illustrates the non-sufficient supply of the primary factor.}
\usage{
Example4.2()
}
\author{
LI Wu <liwu@staff.shu.edu.cn>
}
\references{
LI Wu (2019, ISBN: 9787521804225) General Equilibrium and Structural Dynamics: Perspectives of New Structural Economics. Beijing: Economic Science Press. (In Chinese)
}
