\name{Example4.9}
\alias{Example4.9}
\title{Example 4.9 in Li (2019)}
\description{This is Example 4.9 in Li (2019),which illustrates the price signal.}
\usage{Example4.9()}
\author{
LI Wu <liwu@staff.shu.edu.cn>
}
\references{
LI Wu (2019, ISBN: 9787521804225) General Equilibrium and Structural Dynamics: Perspectives of New Structural Economics. Beijing: Economic Science Press. (In Chinese)
}
