\name{Example5.10}
\alias{Example5.10}
\title{Example 5.10 in Li (2019)}
\description{This is Example 5.10 in Li (2019),which illustrates pollution.}
\usage{Example5.10()}
\author{
LI Wu <liwu@staff.shu.edu.cn>
}
\references{
LI Wu (2019, ISBN: 9787521804225) General Equilibrium and Structural Dynamics: Perspectives of New Structural Economics. Beijing: Economic Science Press. (In Chinese)
}
