\name{Example5.3.1}
\alias{Example5.3.1}
\title{First Part of Example 5.3 in Li (2019)}
\description{This is the first part of Example 5.3 in Li (2019),which illustrates fixed assets.}
\usage{Example5.3.1()}
\author{
LI Wu <liwu@staff.shu.edu.cn>
}
\references{
LI Wu (2019, ISBN: 9787521804225) General Equilibrium and Structural Dynamics: Perspectives of New Structural Economics. Beijing: Economic Science Press. (In Chinese)
}
