\name{Example5.6}
\alias{Example5.6}
\title{Example 5.6 in Li (2019)}
\description{This is Example 5.6 in Li (2019),which illustrates fixed assets.}
\usage{Example5.6()}
\author{
LI Wu <liwu@staff.shu.edu.cn>
}
\references{
LI Wu (2019, ISBN: 9787521804225) General Equilibrium and Structural Dynamics: Perspectives of New Structural Economics. Beijing: Economic Science Press. (In Chinese)
}
