\name{Example6.6.2}
\alias{Example6.6.2}
\title{Second Part of Example 6.6 in Li (2019)}
\description{This is the second part of Example 6.6 in Li (2019),which illustrates the first country of a two-country economy.}
\usage{Example6.6.2()}
\author{
LI Wu <liwu@staff.shu.edu.cn>
}
\references{
LI Wu (2019, ISBN: 9787521804225) General Equilibrium and Structural Dynamics: Perspectives of New Structural Economics. Beijing: Economic Science Press. (In Chinese)
}
