\name{Example6.6.3}
\alias{Example6.6.3}
\title{Third Part of Example 6.6 in Li (2019)}
\description{This is the third part of Example 6.6 in Li (2019),which illustrates the second country of a two-country economy.}
\usage{Example6.6.3()}
\author{
LI Wu <liwu@staff.shu.edu.cn>
}
\references{
LI Wu (2019, ISBN: 9787521804225) General Equilibrium and Structural Dynamics: Perspectives of New Structural Economics. Beijing: Economic Science Press. (In Chinese)
}
