\name{Example7.1}
\alias{Example7.1}
\title{Example 7.1 in Li (2019)}
\description{This is Example 7.1 in Li (2019),which illustrates a monetary pure exchange economy.}
\usage{Example7.1()}
\author{
LI Wu <liwu@staff.shu.edu.cn>
}
\references{
LI Wu (2019, ISBN: 9787521804225) General Equilibrium and Structural Dynamics: Perspectives of New Structural Economics. Beijing: Economic Science Press. (In Chinese)
}
