\name{Example7.10.2}
\alias{Example7.10.2}
\title{Extra Part of Example 7.10 in Li (2019)}
\description{This is an extra part of Example 7.10 in Li (2019), which illustrates fiat money and representative money.}
\usage{Example7.10.2()}
\author{
LI Wu <liwu@staff.shu.edu.cn>
}
\references{
LI Wu (2019, ISBN: 9787521804225) General Equilibrium and Structural Dynamics: Perspectives of New Structural Economics. Beijing: Economic Science Press. (In Chinese)
}
