\name{Example7.11}
\alias{Example7.11}
\title{Example 7.11 in Li (2019)}
\description{This is Example 7.11 in Li (2019), which illustrates bond.}
\usage{Example7.11()}
\author{
LI Wu <liwu@staff.shu.edu.cn>
}
\references{
LI Wu (2019, ISBN: 9787521804225) General Equilibrium and Structural Dynamics: Perspectives of New Structural Economics. Beijing: Economic Science Press. (In Chinese)
}
