\name{Example7.12}
\alias{Example7.12}
\title{Example 7.12 in Li (2019)}
\description{This is Example 7.12 in Li (2019), which illustrates the foreign exchange rate and international credit.}
\usage{Example7.12()}
\author{
LI Wu <liwu@staff.shu.edu.cn>
}
\references{
LI Wu (2019, ISBN: 9787521804225) General Equilibrium and Structural Dynamics: Perspectives of New Structural Economics. Beijing: Economic Science Press. (In Chinese)
}
