\name{Example7.13}
\alias{Example7.13}
\title{Example 7.13 in Li (2019)}
\description{This is Example 7.13 in Li (2019), which illustrates indirect financing based on commercial banks.}
\usage{Example7.13()}
\author{
LI Wu <liwu@staff.shu.edu.cn>
}
\references{
LI Wu (2019, ISBN: 9787521804225) General Equilibrium and Structural Dynamics: Perspectives of New Structural Economics. Beijing: Economic Science Press. (In Chinese)
}
