\name{Example7.14}
\alias{Example7.14}
\title{Example 7.14 in Li (2019)}
\description{This is Example 7.14 in Li (2019), which illustrates shadow prices.}
\usage{Example7.14()}
\author{
LI Wu <liwu@staff.shu.edu.cn>
}
\references{
LI Wu (2019, ISBN: 9787521804225) General Equilibrium and Structural Dynamics: Perspectives of New Structural Economics. Beijing: Economic Science Press. (In Chinese)
}
