\name{Example7.15}
\alias{Example7.15}
\title{Example 7.15 in Li (2019)}
\description{This is Example 7.15 in Li (2019), which illustrates shadow prices and international trade.}
\usage{Example7.15()}
\author{
LI Wu <liwu@staff.shu.edu.cn>
}
\references{
LI Wu (2019, ISBN: 9787521804225) General Equilibrium and Structural Dynamics: Perspectives of New Structural Economics. Beijing: Economic Science Press. (In Chinese)
}
