\name{Example7.3}
\alias{Example7.3}
\title{Example 7.3 in Li (2019)}
\description{This is Example 7.3 in Li (2019),which illustrates a monetary Leontief corn economy.}
\usage{Example7.3()}
\author{
LI Wu <liwu@staff.shu.edu.cn>
}
\references{
LI Wu (2019, ISBN: 9787521804225) General Equilibrium and Structural Dynamics: Perspectives of New Structural Economics. Beijing: Economic Science Press. (In Chinese)
}
