\name{Example7.4}
\alias{Example7.4}
\title{Example 7.4 in Li (2019)}
\description{This is Example 7.4 in Li (2019),which illustrates a monetary Cobb-Douglas positive-growth corn economy.}
\usage{Example7.4()}
\author{
LI Wu <liwu@staff.shu.edu.cn>
}
\references{
LI Wu (2019, ISBN: 9787521804225) General Equilibrium and Structural Dynamics: Perspectives of New Structural Economics. Beijing: Economic Science Press. (In Chinese)
}
