\name{Example7.5.2}
\alias{Example7.5.2}
\title{Second Part of Example 7.5 in Li (2019)}
\description{This is the second part of Example 7.5 in Li (2019), which illustrates a monetary Cobb-Douglas corn economy including dividend.}
\usage{Example7.5.2()}
\author{
LI Wu <liwu@staff.shu.edu.cn>
}
\references{
LI Wu (2019) General Equilibrium and Structural Dynamics: Perspectives of New Structural Economics. Beijing: Economic Science Press. (In Chinese)
}
