\name{Example7.7}
\alias{Example7.7}
\title{Example 7.7 in Li (2019)}
\description{This is Example 7.7 in Li (2019), which illustrates foreign exchange rates.}
\usage{Example7.7()}
\author{
LI Wu <liwu@staff.shu.edu.cn>
}
\references{
LI Wu (2019, ISBN: 9787521804225) General Equilibrium and Structural Dynamics: Perspectives of New Structural Economics. Beijing: Economic Science Press. (In Chinese)
}
