\name{Example7.8}
\alias{Example7.8}
\title{Example 7.8 in Li (2019)}
\description{This is Example 7.8 in Li (2019), which illustrates commodity money.}
\usage{Example7.8()}
\author{
LI Wu <liwu@staff.shu.edu.cn>
}
\references{
LI Wu (2019, ISBN: 9787521804225) General Equilibrium and Structural Dynamics: Perspectives of New Structural Economics. Beijing: Economic Science Press. (In Chinese)
}
