\name{Example7.9X}
\alias{Example7.9X}
\title{Example 7.9 in Li (2019)}
\description{This is Example 7.9 in Li (2019), which illustrates commodity money and representative money.}
\usage{Example7.9X()}
\author{
LI Wu <liwu@staff.shu.edu.cn>
}
\references{
LI Wu (2019, ISBN: 9787521804225) General Equilibrium and Structural Dynamics: Perspectives of New Structural Economics. Beijing: Economic Science Press. (In Chinese)
}
