\name{Example8.2}
\alias{Example8.2}
\title{Example 8.2 in Li (2019)}
\description{This is Example 8.2 in Li (2019), which expounds a Cobb-Douglas market-clearing exchange process.}
\usage{Example8.2()}
\author{
LI Wu <liwu@staff.shu.edu.cn>
}
\references{
LI Wu (2019, ISBN: 9787521804225) General Equilibrium and Structural Dynamics: Perspectives of New Structural Economics. Beijing: Economic Science Press. (In Chinese)
}
